Rename Configuration to Dependencies and sort
