+++ Passepartout/Shared/Dependencies.swift
@@ -1,0 +1,94 @@
+//
+//  Dependencies.swift
+//  Passepartout
+//
+//  Created by Davide De Rosa on 12/2/24.
+//  Copyright (c) 2024 Davide De Rosa. All rights reserved.
+//
+//  https://github.com/passepartoutvpn
+//
+//  This file is part of Passepartout.
+//
+//  Passepartout is free software: you can redistribute it and/or modify
+//  it under the terms of the GNU General Public License as published by
+//  the Free Software Foundation, either version 3 of the License, or
+//  (at your option) any later version.
+//
+//  Passepartout is distributed in the hope that it will be useful,
+//  but WITHOUT ANY WARRANTY; without even the implied warranty of
+//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//  GNU General Public License for more details.
+//
+//  You should have received a copy of the GNU General Public License
+//  along with Passepartout.  If not, see <http://www.gnu.org/licenses/>.
+//
+
+import CommonLibrary
+import CommonUtils
+import Foundation
+import PassepartoutKit
+
+enum Dependencies {
+    enum ExtendedTunnel {
+    }
+
+    enum IAPManager {
+    }
+
+    enum ProfileManager {
+    }
+}
+
+extension Dependencies.IAPManager {
+
+    @MainActor
+    static let inAppHelper = StoreKitHelper(
+        products: AppProduct.all,
+        inAppIdentifier: {
+            let prefix = BundleConfiguration.mainString(for: .iapBundlePrefix)
+            return "\(prefix).\($0.rawValue)"
+        }
+    )
+
+    static var betaChecker: BetaChecker {
+        TestFlightChecker()
+    }
+
+    static let productsAtBuild: BuildProducts<AppProduct> = {
+#if os(iOS)
+        if $0 <= 2016 {
+            return [.Full.iOS]
+        } else if $0 <= 3000 {
+            return [.Features.networkSettings]
+        }
+        return []
+#elseif os(macOS)
+        if $0 <= 3000 {
+            return [.Features.networkSettings]
+        }
+        return []
+#else
+        return []
+#endif
+    }
+}
+
+extension Dependencies.ProfileManager {
+    static let sharedTitle: @Sendable (Profile) -> String = {
+        String(format: Constants.shared.tunnel.profileTitleFormat, $0.name)
+    }
+
+#if os(tvOS)
+    static let mirrorsRemoteRepository = true
+
+    static let isIncluded: @MainActor @Sendable (CommonLibrary.IAPManager, Profile) -> Bool = {
+        $1.attributes.isAvailableForTV == true
+    }
+#else
+    static let mirrorsRemoteRepository = false
+
+    static let isIncluded: @MainActor @Sendable (CommonLibrary.IAPManager, Profile) -> Bool = { _, _ in
+        true
+    }
+#endif
+}
