+++ Passepartout/Shared/Testing/AppContext+Testing.swift
@@ -33,7 +33,7 @@ extension AppContext {
     static func forUITesting(withRegistry registry: Registry) -> AppContext {
         let iapManager = IAPManager(
             customUserLevel: .subscriber,
-            inAppHelper: Configuration.IAPManager.inAppHelper,
+            inAppHelper: Dependencies.IAPManager.inAppHelper,
             receiptReader: FakeAppReceiptReader(),
             betaChecker: TestFlightChecker(),
             productsAtBuild: { _ in
